+++ CHANGELOG.md
@@ -5,10 +5,13 @@
 ---
 
 - Added Spellcheck - [#277](https://github.com/hkdb/aerion/issues/277) - (Unfortunately increases the binary size)
+- Improved profile pic support - [#183](https://github.com/hkdb/issues/183)
+- Added re-auth button to contact write
 - Improved unread count and IDLE sync - [#327](https://github.com/hkdb/aerion/issues/327)
 - True fix for duplicate e-mail add error - [#318](https://github.com/hkdb/aerion/issues/318)
 - Fixed calendar auth on add - [#337](https://github.com/hkdb/aerion/issues/337)
 - Added guard against wails bridge saturation dervied from calendar rapid view switch
+- Fixed CarDAV sync profile pic bug
 - Updated i18n translation